map drawing to move with the curr position, not yet aware of the zoom

+++ src/java/com/astoev/cave/survey/activity/map/MapView.java
@@ -504,11 +504,11 @@ public class MapView extends View {
         return scale;
     }
 
-    public int getMoveX() {
+    public float getMoveX() {
         return mapCenterMoveX;
     }
 
-    public int getMoveY() {
+    public float getMoveY() {
         return mapCenterMoveY;
     }
